Make browserid_info a function to fix the form CSRF.
Django's CSRF protection is difficult to use outside of a view. When
browserid_info was being generated in the context processor, it couldn't
get the {% csrf_token %} tag to output a CSRF token.

To fix this, we change browserid_info to a function that executes when
the template is rendered. By that point, the Django CSRF plumbing has
sorted itself out and {% csrf_token %} now outputs a token correctly.
This also has the side benefit of making the API a bit more consistent.