Issue 138. Fix login failure auto-login
If you have BROWSERID_CREATE_USER = False, then try to log in with a valid
account on Persona, but not on the site, then you end up in this endless
loop of going to a page on the site and being redirected to the login
failure page immediately.

This fixes that.