Ensure backend attribute exists on user objects.
A user's email is only included in the browserid_info div if it was
authenticated by the BrowserIDBackend, which is signified by the
backend attribute on the user object. However, if a user is retrieved
from the session, the backend attribute isn't set by default.

In that case, the user object is retrieved by the get_user method
on the backend that first logged the user in. This commit updates that
method so that it sets the backend attribute on the user object, 
ensuring that it will always have that attribute and can be correctly
identified as being authed by the BrowserIDBackend.

Oh, and I also added Python 3.3 to the allowed failures list in Travis
until we figure out why lxml doesn't want to install on it.