Hey ya'll, it's a release!
Well, okay. I'm incrementing the version number because it's been 5
months and new functionality has been added, but I'm not quite happy
with things, so I'm not going to announce it to the mailing lists yet.

We really need to go through and clean up the API to remove unnecessary
settings, functions, etc. We also need to make JS testing actually 
happen, add real Python 3 support, etc. And it's time we did that
instead of adding new features. 

Onward!