Documentation for using Django-BrowserId with AJAX.
I recently had to piece together how to do this and this is how I did
it. It may not be complete, but it is a good start. One thing I am
uncertain of is what to do with `next`. I had to set it to something
that resolves otherwise, the AJAX handler would attempt to redirect and
on 404, fail which caused the login to fail. It would be better since
this is done via AJAX, to not need next.