Fix docs building
To build the docs, we have to be able to import
django_browserid.tests.settings. To import that, it first imports
django_browserid.tests.__init__ which imports django_browserid.auth
which imports something which requires settings, but that's not loaded
yet, so everything goes down in a firey spiral of doom.

This fixes that by ridiculously importing those two things in the method
they're used rather than at the top level.