Fix #196: Add support for BrowserID in the admin interface.
Also changes browserid_login and browserid_logout to not auto-add the
browserid-login/browserid-logout classes to links that they generate,
in order to allow code to use the helpers without forcing the buttons
to use the login code from browserid.js (the admin utilizes this).

Also also adds some licensing comments that we were missing.