Rework the auto-login docs
I reworked the auto-login docs turning the instructions into lists of
steps and breaking up the text into an "enable" section and a "disable"
section. I think this is easier to read. It's certainly easier to
follow.

Further, I think this addresses the confusion around what happens if
you're not using "django_browserid.helpers.browserid_js" regarding
"browserid/autologin.js" since it explicitly addresses it in the
relevant areas as a discrete step.

Fixes #267